Refactor deployment script
- Move to its own package under deploy/ to reduce
  clutter in repo root
- Move tests inside deploy/ too
- Tests earlier had a recursive dependency - hub.py will call them,
  but they will import hub.py to know more about the hub they are
  testing! This didn't break because the tests were called via
  pytest.main. This did prevent the tests from being used elsewhere.
  The tests now are independent, and get just the info they need
  from commandline arguments
- Temporarily disable the custom home page template test, since I
  haven't quite figured out how to pass the information it
  needs

Fixes https://github.com/2i2c-org/pilot-hubs/issues/139
